backend/requirements.txt
bcrypt==5.0.0
blinker==1.9.0
certifi==2025.10.5
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dotenv==0.9.9
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT==0.3.2
Flask-JWT-Extended==4.7.1
fonttools==4.60.1
itsdangerous==2.2.0
Jinja2==3.1.6
jplephem==2.23
keyboard==0.13.5
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mysql-connector==2.2.9
mysql-connector-python==9.5.0
numpy==2.3.4
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
sgp4==2.25
six==1.17.0
skyfield==1.53
urllib3==2.5.0
weatherapipython @ git+https://github.com/weatherapicom/python.git@05344f2d04cbf0b9513cb69b71f56a66a348e1be
Werkzeug==3.1.3
